Some more tweaks to make it better

+++ convert_to_qcow.sh
@@ -5,5 +5,5 @@ sync
 kpartx -d disk.raw
 sync
 
-rm disk.qcow2
+[ -e disk.qcow2 ] && rm disk.qcow2
 qemu-img convert -f raw -O qcow2 disk.raw disk.qcow2
